created file

+++ python-if_else_loops_functions/1-last_digit.py
@@ -2,7 +2,7 @@
 import random
 number = random.randint(-10000, 10000)
 
-ld = number - (number / 10) * 10
+ld = number - int(number / 10) * 10
 if ld > 5:
     print(f"Last digit of {number} is {ld} and is greater than 5")
 elif ld < 6 and ld != 0:
